# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Night City RP Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /characters:
    get:
      operationId: listMyCharacters
      tags: [characters]
      summary: List the current user's characters (PC + NPC)
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Character" }
    post:
      operationId: createCharacter
      tags: [characters]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CharacterInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Character" }
  /characters/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getCharacter
      tags: [characters]
      responses:
        "200":
          description: Character
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Character" }
    patch:
      operationId: updateCharacter
      tags: [characters]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CharacterUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Character" }
    delete:
      operationId: deleteCharacter
      tags: [characters]
      responses:
        "204": { description: Deleted }
  /characters/{id}/activate:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    post:
      operationId: setActiveCharacter
      tags: [characters]
      summary: Set this character as the user's currently active one
      responses:
        "204": { description: Activated }

  # ---------------- INVENTORY ----------------
  /characters/{id}/inventory:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getCharacterInventory
      tags: [inventory]
      responses:
        "200":
          description: Inventory
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InventoryItem" }
    post:
      operationId: addInventoryItem
      tags: [inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InventoryItemInput" }
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryItem" }
  /characters/{id}/inventory/{itemId}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
      - { in: path, name: itemId, required: true, schema: { type: integer } }
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InventoryItemUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryItem" }
    delete:
      operationId: removeInventoryItem
      tags: [inventory]
      responses:
        "204": { description: Removed }

  # ---------------- WALLET (UnbelievaBoat is source of truth) ----------------
  /characters/{id}/wallet:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getWallet
      tags: [wallet]
      summary: Live balance from UnbelievaBoat
      responses:
        "200":
          description: Wallet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Wallet" }
  /characters/{id}/wallet/transfer:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    post:
      operationId: transferEddies
      tags: [wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransferInput" }
      responses:
        "200":
          description: Transferred
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Wallet" }
  /characters/{id}/wallet/transactions:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getWalletTransactions
      tags: [wallet]
      responses:
        "200":
          description: Transaction history
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WalletTransaction" }

  # ---------------- STATUS (LOA / Attend / Open Shop) ----------------
  /characters/{id}/status:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getCharacterStatus
      tags: [status]
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CharacterStatus" }
    patch:
      operationId: updateCharacterStatus
      tags: [status]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CharacterStatusUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CharacterStatus" }

  # ---------------- PUBLIC DIRECTORIES ----------------
...
  /directory/ripperdocs/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getRipperdocPublic
      tags: [directory]
      responses:
        "200":
          description: Ripperdoc
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RipperdocPublic" }
...
  /directory/stores/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer } }
    get:
      operationId: getStorePublic
      tags: [directory]
      responses:
        "200":
          description: Store
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StorePublic" }

  # ---------------- CATALOGS (read-only public) ----------------
...
  /dice/roll:
    post:
      operationId: rollDice
      tags: [dice]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DiceRollInput" }
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DiceRollResult" }
...
components:
  schemas:
...
    Character:
      type: object
      required: [id, ownerId, name, kind, createdAt]
      properties:
        id: { type: integer }
        ownerId: { type: string }
        name: { type: string }
        kind: { type: string, enum: [pc, npc] }
        archetype: { type: ["string", "null"] }
        background: { type: ["string", "null"] }
        portraitUrl: { type: ["string", "null"] }
        discordChannelId: { type: ["string", "null"] }
        isActive: { type: boolean }
        approved: { type: boolean }
        createdAt: { type: string, format: date-time }
    CharacterInput:
      type: object
      required: [name, kind]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        kind: { type: string, enum: [pc, npc] }
        archetype: { type: string }
        background: { type: string }
        portraitUrl: { type: string }
    CharacterUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        archetype: { type: string }
        background: { type: string }
        portraitUrl: { type: string }
    InventoryItem:
      type: object
      required: [id, characterId, name, quantity]
      properties:
        id: { type: integer }
        characterId: { type: integer }
        name: { type: string }
        category: { type: ["string", "null"] }
        quantity: { type: integer }
        notes: { type: ["string", "null"] }
        equipped: { type: boolean }
    InventoryItemInput:
      type: object
      required: [name, quantity]
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string }
        quantity: { type: integer, minimum: 1 }
        notes: { type: string }
        equipped: { type: boolean }
    InventoryItemUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string }
        quantity: { type: integer, minimum: 0 }
        notes: { type: string }
        equipped: { type: boolean }
    Wallet:
      type: object
      required: [characterId, balance, source]
      properties:
        characterId: { type: integer }
        balance: { type: integer }
        bank: { type: integer }
        cash: { type: integer }
        source: { type: string, enum: [unbelievaboat, local] }
        updatedAt: { type: string, format: date-time }
    TransferInput:
      type: object
      required: [toCharacterId, amount]
      properties:
        toCharacterId: { type: integer }
        amount: { type: integer, minimum: 1 }
        memo: { type: string }
    WalletTransaction:
      type: object
      required: [id, characterId, amount, kind, createdAt]
      properties:
        id: { type: integer }
        characterId: { type: integer }
        amount: { type: integer }
        kind: { type: string, enum: [transfer_in, transfer_out, payout, rent, cyberware, admin, shop, other] }
        memo: { type: ["string", "null"] }
        counterpartyName: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    CharacterStatus:
      type: object
      required: [characterId, loa, attending, openShop, updatedAt]
      properties:
        characterId: { type: integer }
        loa: { type: boolean }
        loaReturnsAt: { type: ["string", "null"], format: date-time }
        attending: { type: boolean }
        openShop: { type: boolean }
        statusMessage: { type: ["string", "null"] }
        updatedAt: { type: string, format: date-time }
    CharacterStatusUpdate:
      type: object
      properties:
        loa: { type: boolean }
        loaReturnsAt: { type: string, format: date-time }
        attending: { type: boolean }
        openShop: { type: boolean }
        statusMessage: { type: string }
    RipperdocPublic:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
        ownerName: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        bannerUrl: { type: ["string", "null"] }
        employeeNames:
          type: array
          items: { type: string }
    StorePublic:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
        ownerName: { type: ["string", "null"] }
        kind: { type: string, enum: [guns, gear, clothing, mixed, other] }
        location: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        bannerUrl: { type: ["string", "null"] }
        employeeNames:
          type: array
          items: { type: string }
...
    DiceRollInput:
      type: object
      required: [expression]
      properties:
        expression: { type: string, minLength: 1, maxLength: 64, description: "e.g. 1d20+5, 4d6kh3" }
        label: { type: string }
        characterId: { type: integer }
    DiceRollResult:
      type: object
      required: [id, expression, total, rolls, createdAt]
      properties:
        id: { type: integer }
        expression: { type: string }
        label: { type: ["string", "null"] }
        characterId: { type: ["integer", "null"] }
        characterName: { type: ["string", "null"] }
        rolls:
          type: array
          items: { type: integer }
        modifier: { type: integer }
        total: { type: integer }
        createdAt: { type: string, format: date-time }
...